updating inventory systems
things like adding the appropriate amount of items if you pick up beyond what you can carry work properly.
coming up with a save/load system is proving to be a bit of a bitch. right now the game attempts to load an inventory and a pouch from two resource files. inside these resource files are an array of dictionaries for each item. inside the dictionaries are keys/values for Name and quantity. the game basically loads this file as the player's starting inventory.
i think before coming up with a save/load system, i need to figure out what happens when you die. if it's done the 'classic' way, you basically get kicked back to the title screen and have to re-load your last save. despite being classic, it's not the best way - gameplay-wise - to handle things, i think.
if it's done in a more modern way, in which you automatically come back to life at the last checkpoint but with some kind of penalty (ie., used items don't come back + XP loss), then "saving" should occur upon death. the issue then becomes a balancing act of incentivising the player to use spells vs. saving ingredients.
if spell ingredients are to be a core aspect of gameplay, you should be able to re-try a failed encounter without losing those used ingredients. on the other hand, if found ingredients are plentiful enough in the level (and respawn on death, for example), permanent loss makes more sense, even if it means the player has to go out of their way slightly to recoup spent ingredients.
i think the next big project for the game should be a death system, which itself is a combination of other systems, which need to be built from the ground up.
player needs spawn points. when you die/restart, you reappear at a spawn point, instead of at your death location. your current spawn point is the last one you visit. need to come up with an interesting concept for spawn points/checkpoints/save points, whatever you want to call them. it could just be the beginning of the last map entered - that sounds like an OK start point. damn this is confusing.