re-vamped pause screen/menu
added a menu to the pause screen. this menu can display stats, inventory items, pouch ingredients, and the default player1 controls.
but buttons don't do anything when pressed. there is some potential functionality to be added:
clicking the "status" button should bring the focus over to the stats, where you should be able to read descriptions of each stat (and maybe more).
clicking the "inventory" button should bring the focus over to the inventory items, where you should be able to read item descriptions.
clicking the "pouch" button should bring the focus over to the pouch ingredients, where you should be able to read ingredient descriptions and see which spells each ingredient is used in.
clicking the "controls" button should bring the focus over to read more detailed descriptions of each action, as well as potentially to re-map/customize controls in the future.