audio code refactor / reworking stamina regen
audio code for the player is quite messy. now the regular sword SFX comes from the swordHitbox AudioStreamPlayer. the "bamboo" and "cursor" SFX (for 0 stamina and notice display popup) now have their own audio stream players to prevent sounds from interrupting each other.
most of the sword SFX are played from the swordhitbox audiostreamplayer. the only exception is that of the shade attack, which begins playing the sound an an irregular moment, and so the call to play the shade swipe (as well as changing its audio bus to get the delay/echo effect) is part of the actual animation resource itself.
this isn't exactly ideal because the way each sound is played is not uniform. however, it works, so i'll probably leave it as is for now and come back to it later if needed.
also changed the way sweating works. sweating only occurs when stamina < 0 and does not remain active once stamina > 0. the reason this was done is as follows:
originally, when you ran out of stamina while charging, the chargeUI would disappear completely and a sweat icon would take it's place, indicating mitigated stamina recovery. this sweat icon, and the sweating status, would remain active until stamina went above 30. the issue was that if stamina went BACK below 0 AFTER the sweating status was applied, the chargeUI would not disappear. this was due to the way the logic was programmed to determine whether to fade in or out the chargeUI, and would not have been able to be fixed without re-programming the entire thing (which honestly wouldn't be a bad idea).
additionally, i thought the sweating mechanic was a bit too confusing to understand intuitively if you didn't already know how it worked, so i decided that sweating should only occur when stamina is depleted.
there is an invisible minimum stamina value, and it is currently -15. that means when you spend stamina beyond 0, that difference must be recovered before the chargeUI shows up again. the recovery rate while sweating, as in, the recover rate while stamina < 0, is 60% of the base stamina recovery rate.
the base stamina regeneration rate has been lowered significantly, though this rate goes up along with the amount of time having not used any stamina. this is to discourage only ever using the last bit of stamina you have: it is more efficient to spend all your stamina at once through various attacks and movement abilities and then recover it all at once, since the regeneration rate increases geometrically if it's not interrupted (attacking, charging, rolling, getting hit and picking up an item count as interruptions).
this stamina rate will only increase if stamina > 0. if stamina is below 0, ie. the Player is sweating, it will need to be recovered until it is above 0 before the stamina recovery rate begins increasing.
all of this is to incentivize not only smart usage of stamina, but also smart recovery.