added wolf's LeapUp/pickup animation
going to try making a universal ingredient pickup that should be able to load whichever ingredient via script. the tricky thing about this would be assigning the appropriate ingredient descriptions. not exactly sure how to go about doing this without putting every ingredient description in the game in the universal script, which doesn't seem logical. but i will make an attempt and see what works and would be most efficient in the long run.