created re-usable ingredient scene/updated drops
ingredient pickups now all load from a single IngredientPickup.tscn scene. upon instantiating an IngredientPickup, an ingredient_name (String) must be assigned. example:

	const IngredientPickup = preload("res://assets/Ingredients/IngredientPickup.tscn")
	var ingredientPickup = IngredientPickup.instance()
	ingredientPickup.ingredient_name = "Rock"
	add_child(ingredientPickup)

... will instantiate a Rock pickup.

when added to a Scene, the IngredientPickup checks the IngredientDatabase and matches the received ingredient_name with the appropriate ingredient Resource. this Resource is assigned to the IngredientPickup's "ingredient" variable, making the Resource's variables (such as its sprite texture & item description) readily available to be assigned to the instanced IngredientPickup:

	var ingredient = check_ingredient(ingredient_name)
	$Sprite.texture = ingredient.texture
	ingredient_description = ingredient.description

the Resource's description variable is similar to a dialog script: an Array of 1 or more Dictionaries, with each dictionary containing the "text" key that determines the text to be written to the dialog box, its paired value.