working on ground layer implementation
collision masks and layers determine what collides with what, what blocks what, etc. the collision LAYER bit is the LAYER on which an object is PLACED, and the collision MASK determines which LAYER an object is LOOKING FOR.
so for example: the Player is on its own collision LAYER, and enemy hitboxes are on the Player's collision MASK, because enemy hitboxes are LOOKING FOR whatever is on the Player LAYER.
doodads are on the World LAYER. all moving bodies are on the World MASK - the World collision layer acts as barriers that bodies on the World mask cannot pass through, which include things like trees, rocks, cliffs, ridges, etc.
in addition to the World layer, 4 new layers/masks have been created: WorldTop (bit 13), WorldUpper (bit 14),  WorldLower (bit 15), and WorldBottom (bit 16).
the way the LayerTest map works is that the collision areas are assigned appropriately to their respective collision layers. there are 2 "transition" areas at the entrance to the valley, which assign the appropriate collision mask to bodies passing through them: when a body passes through the upper area, its mask is set to WorldLower and its z_index set to 0, thus it collides with collision areas on the WorldLower bit, and appears BELOW textures on the 1 z_index (eg., the bridge is on the 1 z_index, so when the Player's z_index is 0, he is drawn behind the bridge); when a body passes through the lower area, its mask is set to WorldUpper and its z_index set to 0, thus it collides with collision areas on the WorldUpper bit, and appears above textures on the 1 z_index (so bodies get drawn ON the bridge).
i had to add an extra z_index argument to blood effects so that when an enemy gets hit, the blood splatters on the z_index texture that the enemy was on when it was hit/killed. the issue here is the Crow: the crow does not look at collision layers, as it flies over world objects whenever it moves. determining which z_index to place its hit effects is proving tricky.
a potential solution is to allow for multiple z_indexes per layer. ie., z_index 0 & 1: bottom layer; z_index 2 & 3: lower layer; z_index 4 & 5: upper layer; and z_index 6 & 7: top layer... this as opposed to how i currently have it set up, which is 1 z-index per layer.
another way i'm considering is to just make an Area2D over specific areas to check for flying enemies. if a flying enemy enters an area - say for example the space above the valley - its z_index is lowered by 1. can't tell which would be better to go with in the long run, but the second method is probably easier to implement so i will try that first.
also item drops don't spawn on their appropriate layer/z_index, yet.