added "flying" physics layer
added a dedicated layer for flying enemies to be used in determining elevation level/z_index. so there will be Area2Ds around things like valleys, which - when a Crow, for example, enters it - will lower entering bodies' z_index, and revert the z_index of exiting bodies.
my previous commit mentioned 2 "potential" solutions but i realize the actual solution is to combine both. the Flying layer and Crow z_index modulation is working, but it's not perfect because feathers & blood can still fly over a layer's boundary and end up being drawn on top of the Player. to clarify...
ISSUE:
when a blood effect is created on one “layer” and moves over to another, its z-index doesn’t change. meaning if a blood drop were to go from a z_index 1 area to a z_index 0 area, the blood’s z_index would still be 1, and if the Player walked over it, the Player’s would be 0, which makes the blood get drawn above in FRONT of the player, when it should ALWAYS be behind.
SOLUTION:
each layer should have 2 (or more) z_index values. so:
WorldBottom: 0-1
WorldLower: 2-3
WorldUpper: 4-5
WorldTop: 6-7
the Player and other bodies are always going to be drawn on the higher z_index of a layer, and the blood/killing effects will always be placed on the lower z_index of a layer.
so, the Player, enemies, items, etc. will only ever be on layers 1, 3, 5, or 7, whereas blood effects will only ever be on layers 0, 2, 4, or 6. if i'm right, this should prevent blood from an enemy killed in a valley to appear on the ground above the valley. i don't think it will solve the issue of blood being drawn above the player, however, so i will think more about this