fixing layers
there are still a number of bugs to be fixed:
i've encountered a wolf leaping over the collision barrier from the upper layer to the lower layer. since it didn't pass through the designated transition point, its z_index and collision masks were fucked. this also happened to the Player once. it's possible that simply making sure the boundaries/collisionboxes that separate layers are thick enough is sufficient to prevent this from happening, but more testing will have to occur.
the blood/death effects not drawing accurately when in between layers is still an issue, but, like above, this would become a non-issue if the boundaries/collision boxes that separate layers are wide enough. i'm just doing testing on a very small area to see what works and what doesn't.