bugfixing layers
the way layers are implemented introduces a number of issues:
enemies that are flying, like the crow, ignore world objects, therefore they ignore the barriers that separate upper/lower floors. when they die/get hit, the objects spawned by the crow need to be assigned the appropriate z-index.
the shade attack has a property in the animation that sets its collision masks - this is to remove the "Enemy" collision mask temporarily, allowing the shade attack to pass through enemies. the issue is that it applies all collision masks at once, which don't include the newly added layer-specific collision layers.
a solution to the first problem is to just make the Crow act like other enemies and follow the rules of all collision masks (no "flying" rule). this could possibly save some headaches down the line, too, though it would make potential gameplay situations less interesting