working on layer behaviour / bugfixes
trying to standardize the way i've handled multiple layers to simplify it, but i'm afraid it's going to be complicated anyway.
had to fix multiple bugs along the way:
one had to do with being unable to queue shade attacks during evasive maneuvers. stats affect things like charge rate now. the "level 2" charge checked to see if the charge_count was EXACTLY equal to stats.max_charge, which would be unlikely if the charge_rate varies at very small amounts, which it now does. the check is now if it is equal or greater than (player line 455)
changed the way the shade attack handles its hitboxes. a bug was encountered when collision masks were reapplied the old way: the reapplication did not take into account the new layer masks, so doing a shade attack would disable layer collision.

i remain unconvinced that this method is the best way to do layer switching the way the project is currently. the reason being the way collision layers/masks interact: when a new feature is added, several others must be modified. there is probably a way to do the foundational code better to allow for new features without the need to continuously modify other parts of the code.
for example: hitbox collision masks look for hitbox collision layers, regardless of z_index/which "floor" they're on. a simple way to override this is probably to check the z_index on hit detection, and return a Miss if they differ. but then remains a bigger problem...
the bodies' collision themselves - while added to one of the four new collision layers for the "floors" - still look for each other, regardless. meaning collision will occur between objects/enemies, despite being on different floors. a solution would be to add the 4 "floors" or collision layers for objects to exist on. this might require a lot of refactoring. , but it might not.
i created a global "floor_change" function for the Transitions to use. it works as follows. when a Transition area is entered, it sends a signal along with a reference to the body that entered it, to the Map's script. depending on if it was an Upper or a Lower transition, the appropriate function is run with the correct values hard-coded.  these values are used to run the global "floor_change" function and apply the appropriate z_index & collision masks to the body.