working on enemy behaviour involving layers
layer transitions, in addition to running the updated change_floor global function, now follow up with the set_enemy_collision_mask() method (global.gd line 147):
compares each of the enemies' z_index values with that of the player, and sets their "Enemies" collision layer to true if they are the same. additionally disables and enables their detection radii, ensuring detection if the player transitions onto a floor where there is an enemy detection area already.
if a player & enemy's z_indexes are different, attacks against each other will always result in a "miss". enemies ignore the player if they are on different z_indexes.
still getting unpredictable behaviour from enemies, occasionally. blood & death effects still don't display properly. it is possible to walk "under" a blood effect if it lands on a higher z-index than the one the player is on.
a potential solution would be to set areas where blood & other effects could potentially land (different from point of origin z_index), and either delete/lower the z_index of the blood/effects that enter that area.
a dirtier solution might be to create a unique canvaslayer strictly for blood effects, and add every blood effect as a child to it, making sure it always displays behind the player/other objects.
an object cannot be on two z_indexes at once. therefore, when a blood drop overlaps into another z_index, display issues will occur. this is extremely tricky.