working on blood effect display when changing layers
dirty solution seems to work but needs more testing.
basically, create an area that covers a location that you would like to check the z_index of entering areas - in this case, the z_index of the blood effects. if the blood z_index is 5, changes it to 3.
need to clean up & organize my collision layers.
also, i realize checking the Player's z_index and applying enemies to the Enemies collision layer depending won't work with multiplayer as the different players will be able to have different z_indexes. will need to refactor enemy collision layer into 4 separate ones - for each floor, or level.