hit/death effects display on proper z_index
added Area2D to determine the fly zone. when a flying object enters this zone, its z_index is checked. if it is 5, it is changed to 3 (and back to 5 on exiting the zone). blood spatters are considered flying, so when they pass into the zone, they get a z_index of 3, or on the ground of the valley.
bug will occasionally occur when enemies break their collision bounds. not sure why this happens but will have to test more to pinpoint exact cause.