crows now able to fly under bridge
the way it works is very finicky so i'm not sure how much practical use this will see in the final game but as a programming challenge it sure blasted my brain.
if a crow lands, it begins looking for "FlyZone" areas. if it is in one (the valley is a FlyZone), it's like it's changing floors. its z_index gets changed from 0 to -2, and a collision mask is applied that looks for FlyZones (this collision mask is also applied on the sides under the bridge, so as to prevent the Crow from clipping through the bridge if it tries to leave the valley from below). 0 is considered over the bridge, and -2 is considered under.
enemies still will only target bodies on the same z_index as them, so if you are in the valley and the crow has not yet landed in the valley, the crow will not aggro you. a crow is considered landed in the valley if its z_index when choosing to land is 0, and it is over the valley (FlyZone).
if the crow begins its flying animation, it checks its own z_index. if it is 0, then the crow is no longer scanning for FlyZone areas. if it is -2, then that means it's still in the valley, aka in the FlyZone area: it will need to actually leave the area before it stops scanning for FlyZones. when it leaves, its z_index reverts to 0.
basically, if a crow lands in the valley (not on the bridge), it's like they changed floors: their z_index gets changed from 0 to -2, and collision is applied under the bridge so as to prevent the crow from clipping through it if it decides to fly away. if they land elsewhere above ground, it still checks for FlyZone, but it won't find any unless it's over the valley. when it takes off again it will stop checking for FlyZones (preventing unintended collision with the bridge from above).