continuing Enemy refactor
since the enemy scripts are so large, i've decided i'd begin by consolidating their common methods into one shared Enemy singleton script. gotta say, i don't know exactly what i'm doing, so i'm just slowly migrating the basic functions over from the 3 enemies' scripts to get the hang of it. i think in an ideal situation, ultimately there would be a generic Enemy scene which loads the appropriate parameters for the enemy (stats, sprites/animations, sounds, AI script) from specific Resources. this is how ingredients work: there is one single Ingredient scene, which populates itself with the appropriate  attributes depending on the name it is given. for example, if you were to load an Ingredient and give it the name Rock, it would load a rock ingredient, along with its description and sprite. again, ideally, the same thing would happen with enemies: load an Enemy scene, give it a name, and then the Enemy scene looks for an Enemy resource with that name and fills itself out, so to speak.
but i'll carry on with the method refactor for now and look into loading enemies from resources once the basic common functions are integrated.