created test spell!!
right now everything is hard-coded, which means things will have to be broken up into components. for example i'm guessing that spells will be resources loaded when the player hits the "spell" button (currently known as the "item" button). the logic will probably be similar in structure to handling inventory items. the player should be able to scroll through their available spells, and once the cast button is pressed, that spell's resource is checked for the appropriate parameters - such as, for example, which animation(s) to play (probably have each spell as a scene, and just load the scene), as well as which ingredients to check for removal.
so hitting G first parses through the player's pouch resource, looking for "Rock" and "Clay" and making sure their quantities are above 0. if all that works out, removes 1 rock/clay, and loads the spell.
when a spell is cast, the spell is loaded as a scene and added as a child to World. the spell's _ready() method puts the player in the ACTION state, which is reserved for specific animations, and during which all Player inputs are ignored. it also turns off the AnimationTree, which would play animations normally depending on dir_vector. then, plays an animation "Cast" on the Player's AnimationPlayer, and a sound effect.
the rest depends on the spell.
first, a timer is necessary if you don't want the spell to come out instantly. the startup for this spell is 0.4s, after which point a flash occurs and a gigantic hitbox begins monitoring for hurtboxes. the spell is deleted when the audio finishes, which is 0.05s later (total time: 0.45s).
added a new check in the Enemy script for if a hitbox detected by a hurtbox is a spell. if it is, skips both hit & crit detection and just deals straight damage. still takes into account enemy defense, though.
also split the actual dealing damage part of the enemy_hurtbox_entered function into its own method. this saves code from having to be duplicated to take into account whether a hitbox is that of a spell's or the sword's. some code is still duplicated, however, but i wanted to think more about how to clean it up properly as opposed to just going with the first quick & dirty method i think of.
also, capped damage at 9999, just for the fun of seeing all the 9999s pop up at once when all the enemies are hit by the spell. since the container that holds the labels that display damage values are children of the enemy itself, and the enemy gets deleted as soon as it explodes, the killing blow damage numbers are deleted quicker as well. i wonder if there's a way around that: maybe reparenting the damageContainer to somewhere else in the sceneTree when an enemy dies, then having the damageContainer delete itself after a brief period or something. that makes sense to me.