refactored more enemy code/bugfixes
ran into some issues refactoring death code:
occasionally, a dead enemy's death effect would not appear. this would happen less than 10% of the time. a confusing issue, but i think it has been solved.
when an enemy dies, a number of things occur. things that are common between enemies (for example, fading their sprite color to red) are in the Enemy script, and things specific to each enemy (for example, if a Crow is airborne, it "falls" 16px before exploding) are still on that particular enemy's respective script.
the "fades" (like fading a sprite's colour from one to another over time, or fading a crow's y-position over time) are controlled by a Tween node on the enemy. however, some tweens are activated on the global Enemy script, and some tweens are activated on the specific enemy's script. before, these tweens were followed by a yield function which would delay the code that followed it until all tweens were completed. i'm guessing that since some tweens were activated inside the specific enemy's script and some from outside, the yield function would occasionally glitch because it THINKS all tweens are completed when they aren't. in other words, it is likely that the enemy became queued for deletion on the same frame or before the enemy death effect is being instantiated, and instead of waiting for the tween to complete before instantiating a death effect, it simply deletes the entire node on the same frame yet before the death effect is added to the game, so the death effect is never seen/heard.
my solution had 2 steps:
1. get rid of all yield(tween) calls. this way, code that was intended to be delayed for a certain period would not depend on tween nodes.
2. replace them with a yield(get_tree().create_timer(0.5), "timeout"), which creates a one-shot timer right then, without relying on any other parts of the program for instruction.
some code is still duplicated this way so i will perhaps need to experiment to see where redundancies can be eliminated. for example, the yield(get_tree().create_timer(0.5), "timeout") call exists on all specific enemies' no_health functions as well as the global Enemy script's no_health function. perhaps next i'll try to get all code to be delayed on enemy death in the same function. moving all enemies' queue_free() method to the Enemy script seems to work, so maybe i'll do item drops next.
also added the screen shake scene to the default camera2d scene, and replaced every camera2d with an instance of this scene, found in assets/System/Camera2D.tscn.
finally, there are still instances where Crows leaving the valley do have their z-index updated back to 0. when a crow lands in the valley, its z-index is changed from 0 to -2, and when it leaves, it is supposed to be changed back from -2 to 0, but occasionally (less than 10% of the time it seems) that doesn't happen, resulting in an un-hittable enemy and one that gets drawn behind everything else in the scene. not sure what the most efficient way of solving this would be. a compromise would be to just simply not have flying enemies change z-indexes, which means being on different elevations would result in that enemy never aggroing, but i don't find that satisfactory.