refactored enemy spawn/despawn code
the enemy script now has functions that handle both the instantiating of enemy spawners when an enemy has left the screen, as well as setting the proper HP values to previously despawned enemies when they re-enter the screen.
also, to mitigate the issue where the crow leaves the valley while remaining on the -2 z_index, flyer enemies will have a different PlayerDetectionZone script than regular enemies. for regular enemies, the behaviour remains the same: if a body (the Player) enters an enemy's PlayerDetectionZone, it checks to see if the z_indexes are the same, and if they are, the enemy detects the player (CHASE state), preventing regular enemies from aggroing the player if they are at a different elevation. for flying enemies, the condition includes if the body (Player) is at the same OR HIGHER z_index. effectively, what this means is that flying enemies can aggro you if you are on an elevation ABOVE them. if they see you this way, the enemy's z_index is set to that of the Player's. hopefully this will mitigate some z_index mismatch in the future, but admittedly it is not the most elegant solution.