implemented new screen shake function / many gameplay adjustments
new screen shake takes a value "trauma", which normally decays over time. trauma is always going to be a value between 0 and 1. the more trauma, the more the screen shakes.
testing it out with the Ability template, and experimenting with queued trauma adjustments. can also change the decay time.
got rid of the old screen shake scene as i like this method better.
also pouch ingredients are erased from the pouch resource when quantity = 0, so they don't show up in the pause menu anymore if an ingredient is depleted.