adding ability resources
both the inventory items and pouch ingredients are stored in databases created at the launch of the program by parsing through their respective folders and adding each resource it finds into the game to be referenced by the player's inventory or pouch. this is done via the GameManager script.
i can probably recycle the logic used for the inventory items and apply it to different spells/abilities. hopefully adding the capability to scroll through spells and have the game be able to read the ability resource to assign the necessary properties will be simple. some important properties include:
SpriteFrames, for icon animation;
type, for formulaType (HEAL, DAMAGE, BUFF, DEBUFF - just placeholders for now and will likely be updated or removed as necessary)
cost. which is a Dictionary holding 2 key/value pairs associated with Ingredient name and quantity needed to cast, and;
potency, which should determine ability strength.