added FormulaDatabase & FormulaBook
... to keep track of formulas in the game and formulas obtained by the Player.
logic works similar to Inventory & Pouch logic, in that the FormulaBook is loaded as a resource and assigned to the player.
formulas are added using the FormulaBook's add_formula function. unlike the others' add_item and add_ingredient functions, this one does not take a quantity.
if a formulabook.tres resource is not found in the user's app_userdata/Action_RPG_1 folder, two formulas - Flash and Heal - are added to the player's formulabook, which is then saved to disk as formulabook.tres, and loaded accordingly if it is found at program start.
this is all backend stuff. the actual casting of the Flash formula is still hard-coded into the player controls. i've created a new scene for the Flash formula and assigned it as a variable in the Flash resource. hopefully i can load the casting of the spell, as well as its other parameters like its ingredient cost, from this resource when casting. after implementing that, the next thing will be to implement changing spells in real-time. but in order to test that properly, i will need to create another ability scene. i've already created a resource for it, Heal.tres. it just needs a scene to hold its behaviour in the game, and to load that scene when it's cast.