formulaUI working
the "previous_item" & "next_item" keys, E & R, now scroll through acquired formulas and assign their icon frames (animated sprite) properly. small issue is that the panel itself is an odd amount of pixels in length (19) and the icons are even (16), so they can't be centre-aligned pixel-perfectly. either make panel size even, or icon size odd to fix this issue.
the actual spell-casting/parsing through spells is not implemented yet, as the Flash spell is still hard-coded to the "item" key (G).