heal spell implemented
tested & working. can switch between spells using the "next_item" or "previous_item" keys (E & R).
the ability scenes are technically independent of the rest of the program. they are simply loaded when a formula is cast by getting the respective scene's filepath, which is loaded from the corresponding ability's formula resource, in which there is a "scene" variable which holds the scene's filepath. my brain is a little mush so it's hard to explain further. having an ability equipped, ie. in the ability box, means that formula's resource is available by getting the player's currently selected formula:
player.formulabook._formulas[formulabook.current_selected_formula]