formula quantity display added / bugfixes
spellcasting was not perfectly implemented as i thought yesterday. a crash would occur when attempting to switch formulas with no formulas in possession using the next_item (R) input, whose advance_selected_formula function didn't take into account if there were 0 formulas in the player's formulabook.
added formula quantity display to the formulaUI. the formulaUI has a new method: update_formula_quantity. additionally, it now receives the "ingredient_quantity_updated" signal. when either the current_selected_formula_changed signal OR the ingredient_quantity_updated signal is received by the formulaUI, runs the update_formula_quantity function.
first, this function gets a reference to the player's current pouch ingredients, stored as an Array of Dictionaries. then, it gets a reference to the ingredients needed by the currently selected formula, and stores their keys (names) in the Array ingredients_needed. then, it does the same thing for the quantities needed and stored them in Array quantity_needed. then two values, a & b, are set to 0.
we then iterate through our ingredients Array, and compare the ingredients_needed to what is inside. if the values match (eg. if ingredients[i].ingredient_reference.name == "Rock" && ingredients_needed[0] == "Rock"), integer 'a' is set to the quotient derived from dividing the quantity in our possession of the ingredient in question with the quantity needed. integer 'b' is set the same way, for the formula's next ingredient. the formula_quantity is updated as whichever of these values, a or b, is lower.
this method is inefficient for two reasons:
1. update_formula_quantity is executed even if the player picks up ingredients that aren't required by the current_selected_formula. the quantity won't update, obviously, but the function is run anyway for no reason because it runs whenever it detects a change in ANY currently held ingredient's quantity.
2. the function runs twice when a formula is used. this is because technically two quantities are getting updated: one for each ingredient used in the formula. thus, two signals are emitted.
there is probably a way to prevent issue 1 from happening by checking if the name of the ingredient whose quantity has changed is in the ingredient_reference.cost.keys() of the current_selected_formula, and if it isn't, return out of the function.
as for issue 2, there probably isn't a simple way of preventing this from happening barring re-writing the way formulas/ingredients are updated.
earlier today, i had tried to place all formulabook code into the pouch code, which resulted in some unexpected problems. the primary issue was that when the formula code was placed in the pouch script, the _formulas array (which normally stores references to formulas held by the player) stored data meant for the _ingredients array. the _ingredients array was updated normally, but when it came to adding formulas at runtime, for some reason the _formulas array wasn't empty; it was simply a copy of the _ingredients Array, which resulted in an error when trying to add a formula to the player's formulabook as when we tried to iterate through the _formulas array (which shouldn't even have occurred because it should have been empty), an error was encountered because it was trying to find a formula_reference except instead it found ingredient_reference. now, this didn't make sense to me whatsoever. nothing was accessing the _formulas array so i have no idea why it populated itself with data meant for _ingredients. the strangest thing was that this only occurred at runtime, when a pouch resource was being created and saved to disk. if the exact same code was executed after the pouch had been created, ie. in-game, the formulas were added successfully and were able to be cast. however there were other issues i encountered so i decided it wasn't worth it and the pouch/formulabook remain as separate resources.