working on magic damage algorithm
i will have to come up with spell leveling / determine the spell level modifier before continuing work on the magic damage algorithm.
there are 3 important components to determining base magic damage:
1. the damage value of the spell;
2. the magic modifier;
3. the spell level.
the magic modifier is determined as such: 1 + (MAG * 1/32)
currently, the algorithm looks like this:
damage = base_damage * (magic_mod*magic_mod) * spell_level_mod
the idea is to keep these 3 important components balanced, but it's hard to do so before each component is finalized.
but, basically, the base damage of a spell should not overpower the magic_mod and spell_level_mod of the player. consequently, the magic_mod should not be able to overpower the base_damage of a spell & the spell_level_mod. and, finally, the spell_level_mod should not overpower the base_damage nor the magic_mod.
maybe some examples to illustrate...
A) a weak spell (one with a low base damage) should never become overpowered, even with a high magic stat and high spell level;
B) a low magic stat should equalize damage if the spell's base damage is high and the spell level low, and the same goes for low base damage and high spell level
C) a spell with high base damage should do only modest damage if the spell level is low and the magic level is modest.
basically, the player's MAG stat will determine whether a spell is weaker than it "should" be, or "stronger" than it should be, but the magic stat alone (or lack thereof) should not be able to make a spell overpowered nor underpowered.