implemented magic stat to work with spells/drop rate; spells level up
values are not final obviously. algorithms are hard to design but the basic functions are working, basically just have to plug in proper/balanced values at this point.
spells also level up. i made spell experience accumulate 5x faster than what it will be in the final game. needs a lot of testing/balancing.
i will need to write a function that can get the name of the formula used and apply the appropriate experience/levelling methods based on that. right now, it is hard coded per spell: 10 lines of code per spell to distribute xp (FormulaStats.gd). with a large number of spells that will be a pain to modify, so i might need a real programmer's help to come up with a function that can do what i want from knowing just the name of the spell alone. it should be extremely simple, i just can't think right now because brain fried.