formulaUI ingredient quantity bugfix
fixed bug where ingredient quantity in the GUI would not update from 0. this was because the pouch only emitted the "ingredient_quantity_updated" signal - which the FormulaUI listens for to update itself - if the ingredient was ALREADY in the pouch. going to run through the add_ingredient function just to refresh myself:

add_ingredient(ingredient_name, quantity)

the function takes two arguments, the ingredient_name as a String and the quantity as an integer. the first thing that happens is to make sure the quantity of the ingredient is greater than 0,  then to check the ingredient database to see if the ingredient exists (the result of this check is stored in a variable "ingredient" that we will use if we don't already currently have any of the ingredient being added).
then we iterate through our _ingredients array, checking the ingredients we currently have in our pouch, and also checking to see if the quantity added + the quantity on hand exceeds the maximum stack size for the ingredient (which is 20 currently), and sets it to the max size if it is. if it doesn't, adds the quantity added to the quantity on hand, then emits the "ingredient_quantity_updated" signal.
if we iterate through our entire _ingredients array without finding one with their inventory_ingredient.ingredient_reference.name not matching ingredient_name (the function's original argument taken as a string), that means we currently do not that ingredient in our inventory, and so we must add it by appending it to the _ingredients array.
we do this by checking if quantity > 0. this is because quantity gets set to 0 if we have the ingredient we are adding already, OR we return out of the function if we exceed our max stack size. if we have a quantity > 0, creates a dictionary for our new ingredient that will be added to the _ingredients array. the first key, ingredient_reference, is set to the "ingredient" variable we set when we checked the ingredient at the beginning of the function. the second, quantity, is set to the quantity. we then append this dict to the _ingredients array, set the quantity of 0, and iterate once again through the _ingredients array, checking for the ingredient we added, to emit the signal that will update the FormulaUI.